pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drowsiness-detection"
version = "0.0.1"
description = "A package for detecting drowsiness using computer vision. This is our class project in Computer Vision class."
authors = [
    { name = "Jirayuwat B.", email = "jirayuwat.dev@gmail.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python",
    "numpy",
    "PyYAML",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "black", "flake8", "isort"]

[tool.setuptools]
packages = ["drowsiness_detection"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--cov=drowsiness_detection --cov-report=term-missing"
testpaths = ["tests"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

